Refactored code into separate files for better structure and maintainabililty
